+++ plugin-gradle/src/main/java/com/diffplug/gradle/spotless/SpotlessPlugin.java
@@ -27,7 +27,7 @@ import com.diffplug.spotless.SpotlessCache;
 
 public class SpotlessPlugin implements Plugin<Project> {
 	static final String SPOTLESS_MODERN = "spotlessModern";
-	static final String VER_GRADLE_MIN = "7.3";
+	static final String VER_GRADLE_MIN = "8.1";
 	static final String VER_GRADLE_MIN_VERSION_FOR_CUSTOM = "8.4";
 	private static final int MINIMUM_JRE = 17;
 
@@ -42,7 +42,7 @@ public class SpotlessPlugin implements Plugin<Project> {
 					+ "https://docs.gradle.org/current/userguide/building_java_projects.html#sec:java_cross_compilation");
 		}
 		// if -PspotlessModern=true, then use the modern stuff instead of the legacy stuff
-		if (project.hasProperty(SPOTLESS_MODERN)) {
+		if (GradleCompat.isPropertyPresent(project, SPOTLESS_MODERN)) {
 			project.getLogger().warn("'spotlessModern' has no effect as of Spotless 5.0, recommend removing it.");
 		}
 		// make sure there's a `clean` and a `check`